+++ SM_Mono.Pre-5/Main.cs
@@ -1,4 +1,6 @@
-﻿using MelonLoader.Support.Preferences;
+﻿using System.Runtime.InteropServices;
+using System.Runtime.CompilerServices;
+using MelonLoader.Support.Preferences;
 using UnityEngine;
 
 namespace MelonLoader.Support
@@ -12,12 +14,15 @@ namespace MelonLoader.Support
         {
             Interface = interface_from;
             string game_version = Application.version;
-            MelonLogger.Msg("Game Version: " + game_version);
-            MelonUtils.SetConsoleTitle(MelonUtils.GetVersionStrWithGameName(game_version));
+            MelonLogger.Msg($"Game Version: {game_version}\n");
+            MelonUtils.SetConsoleTitle(GetVersionStrWithGameName(game_version));
             UnityMappers.RegisterMappers();
 
             Component.Create();
             return new SupportModule_To();
         }
+        [MethodImpl(MethodImplOptions.InternalCall)]
+        [return: MarshalAs(UnmanagedType.LPStr)]
+        private extern static string GetVersionStrWithGameName([MarshalAs(UnmanagedType.LPStr)] string GameVersion = null);
     }
 }